Refactoring: move field.rs to secret_sharing/
