+++ src/lib.rs
@@ -3,7 +3,6 @@
 pub mod chunkscan;
 pub mod cli;
 pub mod error;
-pub mod field;
 pub mod helpers;
 pub mod protocol;
 pub mod secret_sharing;
