+++ src/protocol/check_zero.rs
@@ -1,9 +1,8 @@
 use crate::{
     error::BoxError,
-    field::Field,
     helpers::fabric::Network,
     protocol::{context::ProtocolContext, reveal::reveal, RecordId},
-    secret_sharing::Replicated,
+    secret_sharing::{Field, Replicated},
 };
 
 use serde::{Deserialize, Serialize};
@@ -83,8 +82,8 @@ pub async fn check_zero<F: Field, N: Network>(
 #[cfg(test)]
 pub mod tests {
     use crate::error::BoxError;
-    use crate::field::{Field, Fp31};
     use crate::protocol::{check_zero::check_zero, QueryId, RecordId};
+    use crate::secret_sharing::{Field, Fp31};
     use crate::test_fixture::{logging, make_contexts, make_world, share, TestWorld};
 
     #[tokio::test]
