+++ src/protocol/prss.rs
@@ -1,5 +1,4 @@
-use crate::field::Field;
-use crate::secret_sharing::Replicated;
+use crate::secret_sharing::{Field, Replicated};
 use aes::{
     cipher::{generic_array::GenericArray, BlockEncrypt, KeyInit},
     Aes256,
@@ -342,7 +341,7 @@ impl Generator {
 #[cfg(test)]
 pub mod test {
     use super::{Generator, KeyExchange, SequentialSharedRandomness};
-    use crate::{field::Fp31, protocol::UniqueStepId, test_fixture::make_participants};
+    use crate::{protocol::UniqueStepId, secret_sharing::Fp31, test_fixture::make_participants};
     use rand::{thread_rng, Rng};
     use std::mem::drop;
