+++ src/protocol/reveal.rs
@@ -1,9 +1,7 @@
 use crate::helpers::fabric::Network;
 use crate::protocol::context::ProtocolContext;
-use crate::{
-    error::BoxError, field::Field, helpers::Direction, protocol::RecordId,
-    secret_sharing::Replicated,
-};
+use crate::secret_sharing::{Field, Replicated};
+use crate::{error::BoxError, helpers::Direction, protocol::RecordId};
 use embed_doc_image::embed_doc_image;
 use serde::{Deserialize, Serialize};
 
@@ -58,8 +56,8 @@ mod tests {
     use tokio::try_join;
 
     use crate::{
-        field::Fp31,
         protocol::{reveal::reveal, QueryId, RecordId},
+        secret_sharing::Fp31,
         test_fixture::{make_contexts, make_world, share, TestWorld},
     };
