+++ src/protocol/sort/reshare.rs
@@ -1,7 +1,7 @@
 use crate::helpers::fabric::Network;
+use crate::secret_sharing::Field;
 use crate::{
     error::BoxError,
-    field::Field,
     helpers::{Direction, Identity},
     protocol::{context::ProtocolContext, RecordId},
     secret_sharing::Replicated,
@@ -101,9 +101,9 @@ mod tests {
     use tokio::try_join;
 
     use crate::{
-        field::Fp31,
         helpers::Identity,
         protocol::{sort::reshare::Reshare, QueryId, RecordId},
+        secret_sharing::Fp31,
         test_fixture::{make_contexts, make_world, share, validate_and_reconstruct, TestWorld},
     };
