+++ src/secret_sharing/malicious_replicated.rs
@@ -6,9 +6,8 @@ use std::{
     ops::{Add, Mul, Neg, Sub},
 };
 
-use crate::field::Field;
 use crate::helpers::Identity;
-use crate::secret_sharing::Replicated;
+use crate::secret_sharing::{Field, Replicated};
 
 #[derive(Clone, Copy, PartialEq, Eq)]
 pub struct MaliciousReplicated<F> {
@@ -110,9 +109,8 @@ impl<F: Field> Mul<F> for MaliciousReplicated<F> {
 #[cfg(test)]
 mod tests {
     use super::MaliciousReplicated;
-    use crate::field::Field;
-    use crate::field::Fp31;
     use crate::helpers::Identity;
+    use crate::secret_sharing::{Field, Fp31};
     use crate::test_fixture::{share, validate_and_reconstruct};
     use proptest::prelude::Rng;
