+++ src/secret_sharing/mod.rs
@@ -1,6 +1,8 @@
+mod field;
 mod malicious_replicated;
 mod replicated;
 mod shamir;
 
+pub use field::{Field, Fp31};
 pub use replicated::Replicated;
 pub use shamir::Shamir;
