+++ src/test_fixture/mod.rs
@@ -7,9 +7,9 @@ mod world;
 use self::fabric::InMemoryEndpoint;
 use crate::helpers::Identity;
 use crate::protocol::context::ProtocolContext;
+use crate::protocol::prss::Endpoint as PrssEndpoint;
 use crate::protocol::Step;
-use crate::secret_sharing::Replicated;
-use crate::{field::Fp31, protocol::prss::Endpoint as PrssEndpoint};
+use crate::secret_sharing::{Fp31, Replicated};
 use rand::rngs::mock::StepRng;
 use rand::thread_rng;
 use std::sync::Arc;
